+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.16.2] - 2021-09-19
+
+### Added
+
+-   Added status channel permission set (requires the bot to have manage channel permission)
+
+### Changed
+
+-   The bot will now remove all status messages and send new ones at launch
+
+### Fixed
+
+-   Probably fixed status message being sent again (new system)
+
 ## [1.16.1] - 2021-09-11
 
 ### Added